# Unify `InMemoryStorage` and `_CachedStorage`.