--- a/optuna/storages/_in_memory.py
+++ b/optuna/storages/_in_memory.py
@@ -12,8 +12,8 @@ import uuid
 
 from optuna import distributions  # NOQA
 from optuna.exceptions import DuplicatedStudyError
-from optuna.storages._base import DEFAULT_STUDY_NAME_PREFIX
 from optuna.storages._base import _BackEnd
+from optuna.storages._base import DEFAULT_STUDY_NAME_PREFIX
 from optuna.storages._cached_storage import _CachedStorage
 from optuna.study import StudyDirection
 from optuna.study import StudySummary