## Motivation
- `InMemoryStorage` and `_CachedStorage` share many functionalities. Unifying them will increase code-maintainability.
- Using `_CachedStorage` in all backends make it easy to add further caching functionalities e.g., sampler-specific caches.

## Description of the changes
Unify `InMemoryStorage` and `_CachedStorage`.

We can simplify `InMemoryStorage` more aggressively, but I didn't apply the change to make this PR smaller.

## Microbenchmark
I compared the optimization speed of the current master and this PR with the following code.

<details><summary>Code</summary>
<p>

```python
import time
import optuna


def objective(trial):
    return sum(trial.suggest_float(f"param{i}", 0.0, 1.0) for i in range(3))


if __name__ == "__main__":
    optuna.logging.set_verbosity(optuna.logging.WARNING)
    start = time.time()
    optuna.create_study().optimize(objective, n_trials=5000)
    print(time.time() - start)
```

</p>
</details>

### Result:
| Master  | This PR  | 
|---|---|
|  146.5 sec | 149.8 sec  |

## Notes
This PR depends on #1570.